refactor: Member 디렉토리 추가 및 이동
